+++ packages/playground/src/data/saga.js
@@ -1,15 +1,15 @@
 import { fork } from 'redux-saga/effects';
 
 import setPonyfill from './sagas/setPonyfill';
-import setSpeechRecognitionInstance from './sagas/setSpeechRecognitionInstance';
-import setSpeechSynthesisVoices from './sagas/setSpeechSynthesisVoices';
-import speakUtterance from './sagas/speakUtterance';
-import startSpeechRecognition from './sagas/startSpeechRecognition';
+import speechRecognitionSetInstance from './sagas/speechRecognitionSetInstance';
+import speechRecognitionStart from './sagas/speechRecognitionStart';
+import speechSynthesisSetVoices from './sagas/speechSynthesisSetVoices';
+import speechSynthesisSpeakUtterance from './sagas/speechSynthesisSpeakUtterance';
 
 export default function* () {
   yield fork(setPonyfill);
-  yield fork(setSpeechRecognitionInstance);
-  yield fork(setSpeechSynthesisVoices);
-  yield fork(speakUtterance);
-  yield fork(startSpeechRecognition);
+  yield fork(speechRecognitionSetInstance);
+  yield fork(speechRecognitionStart);
+  yield fork(speechSynthesisSetVoices);
+  yield fork(speechSynthesisSpeakUtterance);
 }
